+++ M2/Front-end/to-do-list/src/js/pages/login.js
@@ -1,4 +1,4 @@
-import { bindLoginEvents } from "../ui/events.js";
+import { bindLoginEvents } from "@events/login/loginEvents.js";
 
 export const initLoginPage = () => {
   bindLoginEvents();
