+++ M2/Front-end/to-do-list/src/js/pages/signup.js
@@ -1,4 +1,4 @@
-import { bindSignupEvents } from "../ui/events.js";
+import { bindSignupEvents } from "@events/signup/signupEvents.js";
 
 export const initSignupPage = () => {
   bindSignupEvents();
