+++ M2/Front-end/to-do-list/src/js/pages/to_do.js
@@ -1,31 +1,33 @@
-import { bindToDoEvents } from "../ui/events.js";
-import { checkSession } from "../services/authService.js";
-import { getCookie } from "../utils/cookie.js";
-import { getUser } from "../services/userService.js";
-import { renderUserName, renderTask, renderTaskStats} from "../ui/render.js";
-import { getTasks } from "../services/taskService.js";
-import { setTasksState } from "../state/taskState.js";
+import { bindToDoEvents } from "@events/todo/todoInit.js";
+import { checkSession } from "@services/authService.js";
+import { getCookie } from "@utils/cookie.js";
+import { getUser } from "@services/userService.js";
+import { getTasks } from "@services/taskService.js";
+import { setTasksState } from "@state/taskState.js";
+import { renderTaskStats } from "@render/taskStatsRender.js";
+import { renderTask } from "@render/taskRender.js";
+import { renderUserName } from "@render/userRender.js";
 
 export const initToDoListPage = async () => {
   const uid = getCookie("uid");
   if (!uid) {
-    window.location.replace("/M2/Front-end/to-do-list/src/pages/login.html");
+    window.location.replace("/src/pages/login.html");
   }
   const isValid = checkSession(uid);
   if (!isValid) {
-    window.location.replace("/M2/Front-end/to-do-list/src/pages/login.html");
+    window.location.replace("/src/pages/login.html");
   }
-  const user = await getUser(uid)
+  const user = await getUser(uid);
   if (user?.name) {
-    renderUserName(user?.name)
+    renderUserName(user?.name);
   }
-  const tasks = await getTasks()
-  setTasksState(tasks)
+  const tasks = await getTasks();
+  setTasksState(tasks);
   if (Array.isArray(tasks) && tasks.length > 0) {
-    tasks.forEach(task => renderTask(task))
-  }else{
-    document.getElementById("not-tasks").hidden=false;
+    tasks.forEach((task) => renderTask(task));
+  } else {
+    document.getElementById("not-tasks").hidden = false;
   }
   bindToDoEvents();
-  renderTaskStats()
+  renderTaskStats();
 };
