+++ M2/Front-end/to-do-list/src/js/ui/events/login/loginEvents.js
@@ -1,0 +1,14 @@
+import { login } from "@services/authService.js";
+
+export const bindLoginEvents = () => {
+  const formLogin = document.getElementById("form-login");
+  if (!formLogin) return;
+  formLogin.addEventListener("submit", async (e) => {
+    e.preventDefault();
+    const success = await login(e.target.uid.value, e.target.password.value);
+    if (success) {
+      location.replace("/src/pages/to-do.html");
+    }
+    return;
+  });
+};
