+++ M2/Front-end/to-do-list/src/js/ui/events/todo/logoutEvents.js
@@ -1,0 +1,11 @@
+import { logout } from "@services/authService.js";
+
+export const bindLogoutEvents = () => {
+  const btnOut = document.getElementById("btn-logout");
+  if (!btnOut) return;
+
+  btnOut.addEventListener("click", () => {
+    logout();
+    location.replace("/src/pages/login.html");
+  });
+};
