+++ M2/Front-end/to-do-list/src/js/ui/render/btnFiltersRender.js
@@ -1,0 +1,8 @@
+export const changeBtnState = (containerFilters, btnClick) => {
+  containerFilters.querySelectorAll("button").forEach((button) => {
+    button.classList.remove("btn-active");
+    button.disabled = false;
+  });
+  btnClick.classList.add("btn-active");
+  btnClick.disabled = true;
+};
