+++ M2/Front-end/to-do-list/src/js/ui/render/userRender.js
@@ -1,0 +1,5 @@
+export const renderUserName = (name) => {
+  const display = document.getElementById("name-log");
+  if (!display) return;
+  display.textContent = name;
+};
