+++ M2/Front-end/to-do-list/src/pages/login.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
@@ -15,7 +15,9 @@
     <div class="displayMsg"></div>
     <main>
       <h1 class="main-title">Login</h1>
-      <p>Don't have an account? <a class="link" href="./signup.html">Sign up</a></p>
+      <p>
+        Don't have an account? <a class="link" href="./signup.html">Sign up</a>
+      </p>
       <form id="form-login">
         <input
           type="text"
@@ -43,7 +45,23 @@
         alt="To Do List"
       />
     </aside>
+    <script type="importmap">
+      {
+        "imports": {
+          "@api/": "/src/js/api/",
+          "@services/": "/src/js/services/",
+          "@state/": "/src/js/state/",
+          "@storage/": "/src/js/storage/",
+          "@utils/": "/src/js/utils/",
+          "@ui/": "/src/js/ui/",
+          "@events/": "/src/js/ui/events/",
+          "@render/": "/src/js/ui/render/",
+          "@pages/": "/src/js/pages/"
+        }
+      }
+    </script>
+
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script type="module" src="../js/main.js"></script>
+    <script type="module" src="/src/js/main.js"></script>
   </body>
 </html>
